requirements.txt
fastapi
uvicorn
httpx
jinja2
aiofiles

starlette
pydantic
python-dotenv